Add ducts and chutes to interact with the high oven controller's inventory
Controller has at most one side available for item piping, which is a bit of a pain to reach. Plus, steelworks had this block
Main difference from steelworks is reliance on inventory slot filters for item insertion. Basically, instead of selecting one of the five input types, 4 input types have a strict filter and the 5 a separate block